# packages/api-contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "BillAssessment": {
        "properties": {
          "approval": {
            "const": "not_performed",
            "default": "not_performed",
            "title": "Approval",
            "type": "string"
          },
          "bill_id": {
            "format": "uuid",
            "title": "Bill Id",
            "type": "string"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/BillCheck"
            },
            "title": "Checks",
            "type": "array"
          },
          "corrections": {
            "items": {
              "$ref": "#/components/schemas/BillCorrection"
            },
            "title": "Corrections",
            "type": "array"
          },
          "corrections_has_more": {
            "title": "Corrections Has More",
            "type": "boolean"
          },
          "detail": {
            "$ref": "#/components/schemas/BillItemsOut"
          },
          "home_id": {
            "format": "uuid",
            "title": "Home Id",
            "type": "string"
          },
          "provenance": {
            "$ref": "#/components/schemas/BillProvenance"
          },
          "read_only": {
            "const": true,
            "default": true,
            "title": "Read Only",
            "type": "boolean"
          },
          "status": {
            "enum": [
              "consistent",
              "warnings",
              "incomplete"
            ],
            "title": "Status",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "home_id",
          "bill_id",
          "status",
          "checks",
          "warnings",
          "provenance",
          "corrections",
          "corrections_has_more",
          "detail"
        ],
        "title": "BillAssessment",
        "type": "object"
      },
      "BillCheck": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "observed": {
            "additionalProperties": true,
            "title": "Observed",
            "type": "object"
          },
          "status": {
            "enum": [
              "pass",
              "warning",
              "unavailable"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "code",
          "status",
          "observed"
        ],
        "title": "BillCheck",
        "type": "object"
      },
...
      "BillCreate": {
        "additionalProperties": false,
        "properties": {
          "amount_dop": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Amount Dop"
          },
          "days": {
            "maximum": 366.0,
            "minimum": 0.0,
            "title": "Days",
            "type": "integer"
          },
          "kwh": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Kwh"
          },
          "period_end": {
            "format": "date",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date",
            "title": "Period Start",
            "type": "string"
          },
          "reading_current": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Current"
          },
          "reading_previous": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Previous"
          },
          "source": {
            "const": "manual",
            "default": "manual",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "period_start",
          "period_end",
          "kwh",
          "amount_dop",
          "days"
        ],
        "title": "BillCreate",
        "type": "object"
      },
      "BillItemIn": {
        "additionalProperties": false,
        "properties": {
          "amount_dop": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Amount Dop"
          },
          "kind": {
            "enum": [
              "charge",
              "discount"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Label",
            "type": "string"
          }
        },
        "required": [
          "label",
          "kind",
          "amount_dop"
        ],
        "title": "BillItemIn",
        "type": "object"
      },
      "BillItemOut": {
        "properties": {
          "amount_dop": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount Dop",
            "type": "string"
          },
          "kind": {
            "enum": [
              "charge",
              "discount"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "position": {
            "title": "Position",
            "type": "integer"
          }
        },
        "required": [
          "position",
          "label",
          "kind",
          "amount_dop"
        ],
        "title": "BillItemOut",
        "type": "object"
      },
      "BillItemsOut": {
        "properties": {
          "bill_amount_dop": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Bill Amount Dop",
            "type": "string"
          },
          "bill_id": {
            "format": "uuid",
            "title": "Bill Id",
            "type": "string"
          },
          "difference_dop": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difference Dop"
          },
          "home_id": {
            "format": "uuid",
            "title": "Home Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/BillItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "items_total_dop": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Items Total Dop"
          }
        },
        "required": [
          "home_id",
          "bill_id",
          "items",
          "items_total_dop",
          "bill_amount_dop",
          "difference_dop"
        ],
        "title": "BillItemsOut",
        "type": "object"
      },
      "BillItemsReplace": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BillItemIn"
            },
            "maxItems": 100,
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "BillItemsReplace",
        "type": "object"
      },
      "BillOut": {
        "properties": {
          "amount_dop": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount Dop",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "days": {
            "title": "Days",
            "type": "integer"
          },
          "home_id": {
            "format": "uuid",
            "title": "Home Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kwh": {
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Kwh",
            "type": "string"
          },
          "period_end": {
            "format": "date",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date",
            "title": "Period Start",
            "type": "string"
          },
          "reading_current": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Current"
          },
          "reading_previous": {
            "anyOf": [
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Previous"
          },
          "source": {
            "enum": [
              "manual",
              "seed"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "id",
          "home_id",
          "period_start",
          "period_end",
          "kwh",
          "amount_dop",
          "days",
          "reading_previous",
          "reading_current",
          "source",
          "created_at"
        ],
        "title": "BillOut",
        "type": "object"
      },
...
      "BillUpdate": {
        "additionalProperties": false,
        "description": "Reemplazo completo (PUT) de una factura manual.",
        "properties": {
          "amount_dop": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Amount Dop"
          },
          "days": {
            "maximum": 366.0,
            "minimum": 0.0,
            "title": "Days",
            "type": "integer"
          },
          "kwh": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              }
            ],
            "title": "Kwh"
          },
          "period_end": {
            "format": "date",
            "title": "Period End",
            "type": "string"
          },
          "period_start": {
            "format": "date",
            "title": "Period Start",
            "type": "string"
          },
          "reading_current": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Current"
          },
          "reading_previous": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,10}|(?=[\\d.]{1,13}0*$)\\d{0,10}\\.\\d{0,2}0*$)",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Previous"
          }
        },
        "required": [
          "period_start",
          "period_end",
          "kwh",
          "amount_dop",
          "days"
        ],
        "title": "BillUpdate",
        "type": "object"
      },
      "BillValidationRequest": {
        "additionalProperties": false,
        "properties": {},
        "title": "BillValidationRequest",
        "type": "object"
      },
      "Body_ocr_bill_photo_api_v1_homes__home_id__bills_ocr_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_ocr_bill_photo_api_v1_homes__home_id__bills_ocr_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OcrDraft": {
        "properties": {
          "amount_dop": {
            "$ref": "#/components/schemas/OcrField"
          },
          "days": {
            "$ref": "#/components/schemas/OcrField"
          },
          "kwh": {
            "$ref": "#/components/schemas/OcrField"
          },
          "period_end": {
            "$ref": "#/components/schemas/OcrField"
          },
          "period_start": {
            "$ref": "#/components/schemas/OcrField"
          },
          "raw_text_excerpt": {
            "title": "Raw Text Excerpt",
            "type": "string"
          },
          "reading_current": {
            "$ref": "#/components/schemas/OcrField"
          },
          "reading_previous": {
            "$ref": "#/components/schemas/OcrField"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "period_start",
          "period_end",
          "days",
          "kwh",
          "amount_dop",
          "reading_previous",
          "reading_current",
          "warnings",
          "raw_text_excerpt"
        ],
        "title": "OcrDraft",
        "type": "object"
      },
      "OcrField": {
        "properties": {
          "confidence": {
            "enum": [
              "high",
              "inferred",
              "none"
            ],
            "title": "Confidence",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "value",
          "confidence"
        ],
        "title": "OcrField",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API para Energy RD - Plataforma de Inteligencia Energ\u00e9tica",
    "title": "Energy RD API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/homes/{home_id}/bills": {
      "get": {
        "operationId": "list_bills_api_v1_homes__home_id__bills_get",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/BillOut"
                  },
                  "title": "Response List Bills Api V1 Homes  Home Id  Bills Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Bills",
        "tags": [
          "bills"
        ]
      },
      "post": {
        "operationId": "create_bill_api_v1_homes__home_id__bills_post",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Bill",
        "tags": [
          "bills"
        ]
      }
    },
    "/api/v1/homes/{home_id}/bills/ocr": {
      "post": {
        "description": "ERD-OCR-01/02: lee una foto de factura y devuelve un BORRADOR para que la persona lo confirme\no corrija. Nunca crea ni modifica una factura: el \u00fanico camino real a la base de datos sigue\nsiendo POST /homes/{home_id}/bills (ya validado), llamado despu\u00e9s de que alguien confirme.",
        "operationId": "ocr_bill_photo_api_v1_homes__home_id__bills_ocr_post",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ocr_bill_photo_api_v1_homes__home_id__bills_ocr_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OcrDraft"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ocr Bill Photo",
        "tags": [
          "bills"
        ]
      }
    },
    "/api/v1/homes/{home_id}/bills/{bill_id}": {
      "delete": {
        "operationId": "delete_bill_api_v1_homes__home_id__bills__bill_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Bill",
        "tags": [
          "bills"
        ]
      },
      "get": {
        "operationId": "get_bill_api_v1_homes__home_id__bills__bill_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Bill",
        "tags": [
          "bills"
        ]
      },
      "put": {
        "operationId": "update_bill_api_v1_homes__home_id__bills__bill_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bill Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Bill",
        "tags": [
          "bills"
        ]
      }
    },
    "/api/v1/homes/{home_id}/bills/{bill_id}/items": {
      "get": {
        "operationId": "get_items_api_v1_homes__home_id__bills__bill_id__items_get",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bill Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillItemsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Items",
        "tags": [
          "bills"
        ]
      },
      "put": {
        "operationId": "replace_items_api_v1_homes__home_id__bills__bill_id__items_put",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bill Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillItemsReplace"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillItemsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Replace Items",
        "tags": [
          "bills"
        ]
      }
    },
    "/api/v1/homes/{home_id}/bills/{bill_id}/validate": {
      "post": {
        "operationId": "validate_bill_api_v1_homes__home_id__bills__bill_id__validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "home_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "bill_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Bill Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BillValidationRequest",
                "default": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BillAssessment"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Validate Bill",
        "tags": [
          "bills"
        ]
      }
    },
...
  }
}